pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "starvector"
version = "1.0"
description = "Generating Scalable Vector Graphics Code from Images and Text"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "accelerate",
    "markdown2[all]",
    "scikit-learn==1.2.2",
    "requests",
    "httpx",
    "uvicorn",
    "fastapi",
    "svgpathtools==1.6.1",
    "seaborn==0.12.2",
    "taming-transformers",
    "lpips",
    "cairosvg",
    "beautifulsoup4",
    "webcolors",
    "omegaconf",
    "open-clip-torch",
    "noise",
    "datasets",
    "scikit-image",
    "fairscale",
    "lxml",
    "clip-openai",
    "sentence-transformers",
    "reportlab",
    "svglib",
    "Pillow",
    "protobuf",
    "openai",
]

[project.optional-dependencies]
train = ["deepspeed", "ninja", "wandb"]

[project.urls]
"Homepage" = "https://starvector.github.io"
"Bug Tracker" = "https://github.com/joanrod/starvector/issues"

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]
